+++ src/Abp.AspNetCore/AspNetCore/MultiTenancy/HttpCookieTenantResolveContributor.cs
@@ -19,7 +19,7 @@ namespace Abp.AspNetCore.MultiTenancy
             _multiTenancyConfig = multiTenancyConfig;
         }
 
-        public int? ResolveTenantId()
+        public long? ResolveTenantId()
         {
             var httpContext = _httpContextAccessor.HttpContext;
             if (httpContext == null)
@@ -33,7 +33,7 @@ namespace Abp.AspNetCore.MultiTenancy
                 return null;
             }
 
-            return int.TryParse(tenantIdValue, out var tenantId) ? tenantId : (int?) null;
+            return long.TryParse(tenantIdValue, out var tenantId) ? tenantId : (long?) null;
         }
     }
 }
