+++ src/Abp.AspNetCore/AspNetCore/MultiTenancy/HttpHeaderTenantResolveContributor.cs
@@ -25,7 +25,7 @@ namespace Abp.AspNetCore.MultiTenancy
             Logger = NullLogger.Instance;
         }
 
-        public int? ResolveTenantId()
+        public long? ResolveTenantId()
         {
             var httpContext = _httpContextAccessor.HttpContext;
             if (httpContext == null)
@@ -46,7 +46,7 @@ namespace Abp.AspNetCore.MultiTenancy
                     );
             }
 
-            return int.TryParse(tenantIdHeader.First(), out var tenantId) ? tenantId : (int?) null;
+            return long.TryParse(tenantIdHeader.First(), out var tenantId) ? tenantId : (long?) null;
         }
     }
 }
