+++ src/Abp.Zero.Common/Authorization/PermissionSetting.cs
@@ -16,7 +16,7 @@ namespace Abp.Authorization
         /// </summary>
         public const int MaxNameLength = 128;
 
-        public virtual int? TenantId { get; set; }
+        public virtual long? TenantId { get; set; }
 
         /// <summary>
         /// Unique name of the permission.
