+++ src/Abp.Zero.Common/Authorization/Users/UserLogin.cs
@@ -20,7 +20,7 @@ namespace Abp.Authorization.Users
         /// </summary>
         public const int MaxProviderKeyLength = 256;
 
-        public virtual int? TenantId { get; set; }
+        public virtual long? TenantId { get; set; }
 
         /// <summary>
         /// Id of the User.
@@ -46,7 +46,7 @@ namespace Abp.Authorization.Users
             
         }
 
-        public UserLogin(int? tenantId, long userId, string loginProvider, string providerKey)
+        public UserLogin(long? tenantId, long userId, string loginProvider, string providerKey)
         {
             TenantId = tenantId;
             UserId = userId;
