+++ src/Abp.Zero.Common/Authorization/Users/UserLoginAttempt.cs
@@ -42,7 +42,7 @@ namespace Abp.Authorization.Users
         /// <summary>
         /// Tenant's Id, if <see cref="TenancyName"/> was a valid tenant name.
         /// </summary>
-        public virtual int? TenantId { get; set; }
+        public virtual long? TenantId { get; set; }
 
         /// <summary>
         /// Tenancy name.
