+++ src/Abp.Zero.Common/Authorization/Users/UserOrganizationUnit.cs
@@ -14,7 +14,7 @@ namespace Abp.Authorization.Users
         /// <summary>
         /// TenantId of this entity.
         /// </summary>
-        public virtual int? TenantId { get; set; }
+        public virtual long? TenantId { get; set; }
 
         /// <summary>
         /// Id of the User.
@@ -45,7 +45,7 @@ namespace Abp.Authorization.Users
         /// <param name="tenantId">TenantId</param>
         /// <param name="userId">Id of the User.</param>
         /// <param name="organizationUnitId">Id of the <see cref="OrganizationUnit"/>.</param>
-        public UserOrganizationUnit(int? tenantId, long userId, long organizationUnitId)
+        public UserOrganizationUnit(long? tenantId, long userId, long organizationUnitId)
         {
             TenantId = tenantId;
             UserId = userId;
