+++ src/Abp.Zero.Common/Localization/ApplicationLanguageTextManager.cs
@@ -39,7 +39,7 @@ namespace Abp.Localization
         /// <param name="culture">Culture</param>
         /// <param name="key">Localization key</param>
         /// <param name="tryDefaults">True: fallbacks to default languages if can not find in given culture</param>
-        public string GetStringOrNull(int? tenantId, string sourceName, CultureInfo culture, string key, bool tryDefaults = true)
+        public string GetStringOrNull(long? tenantId, string sourceName, CultureInfo culture, string key, bool tryDefaults = true)
         {
             var source = _localizationManager.GetSource(sourceName);
 
@@ -53,7 +53,7 @@ namespace Abp.Localization
                 .GetStringOrNull(tenantId, key, culture, tryDefaults);
         }
 
-        public List<string> GetStringsOrNull(int? tenantId, string sourceName, CultureInfo culture, List<string> keys, bool tryDefaults = true)
+        public List<string> GetStringsOrNull(long? tenantId, string sourceName, CultureInfo culture, List<string> keys, bool tryDefaults = true)
         {
             var source = _localizationManager.GetSource(sourceName);
 
@@ -76,7 +76,7 @@ namespace Abp.Localization
         /// <param name="key">Localization key</param>
         /// <param name="value">New localized value.</param>
         [UnitOfWork]
-        public virtual async Task UpdateStringAsync(int? tenantId, string sourceName, CultureInfo culture, string key, string value)
+        public virtual async Task UpdateStringAsync(long? tenantId, string sourceName, CultureInfo culture, string key, string value)
         {
             using (_unitOfWorkManager.Current.SetTenantId(tenantId))
             {
