+++ src/Abp.Zero/Authorization/Users/AbpUserManager.cs
@@ -595,7 +595,7 @@ namespace Abp.Authorization.Users
                 );
         }
 
-        private async Task CheckMaxUserOrganizationUnitMembershipCountAsync(int? tenantId, int requestedCount)
+        private async Task CheckMaxUserOrganizationUnitMembershipCountAsync(long? tenantId, int requestedCount)
         {
             var maxCount = await _organizationUnitSettings.GetMaxUserMembershipCountAsync(tenantId);
             if (requestedCount > maxCount)
@@ -675,7 +675,7 @@ namespace Abp.Authorization.Users
             }
         }
 
-        public virtual void RegisterTwoFactorProviders(int? tenantId)
+        public virtual void RegisterTwoFactorProviders(long? tenantId)
         {
             TwoFactorProviders.Clear();
 
@@ -710,7 +710,7 @@ namespace Abp.Authorization.Users
             }
         }
 
-        public virtual void InitializeLockoutSettings(int? tenantId)
+        public virtual void InitializeLockoutSettings(long? tenantId)
         {
             UserLockoutEnabledByDefault = IsTrue(AbpZeroSettingNames.UserManagement.UserLockOut.IsEnabled, tenantId);
             DefaultAccountLockoutTimeSpan = TimeSpan.FromSeconds(GetSettingValue<int>(AbpZeroSettingNames.UserManagement.UserLockOut.DefaultAccountLockoutSeconds, tenantId));
@@ -826,12 +826,12 @@ namespace Abp.Authorization.Users
             });
         }
 
-        private bool IsTrue(string settingName, int? tenantId)
+        private bool IsTrue(string settingName, long? tenantId)
         {
             return GetSettingValue<bool>(settingName, tenantId);
         }
 
-        private T GetSettingValue<T>(string settingName, int? tenantId) where T : struct
+        private T GetSettingValue<T>(string settingName, long? tenantId) where T : struct
         {
             return tenantId == null
                 ? _settingManager.GetSettingValueForApplication<T>(settingName)
@@ -848,7 +848,7 @@ namespace Abp.Authorization.Users
             return LocalizationManager.GetString(LocalizationSourceName, name, cultureInfo);
         }
 
-        private int? GetCurrentTenantId()
+        private long? GetCurrentTenantId()
         {
             if (_unitOfWorkManager.Current != null)
             {
