+++ src/Abp.ZeroCore.EntityFramework/Zero/EntityFramework/AbpZeroDbMigrator.cs
@@ -47,7 +47,7 @@ namespace Abp.Zero.EntityFramework
         protected virtual void CreateOrMigrate(AbpTenantBase tenant)
         {
             var args = new DbPerTenantConnectionStringResolveArgs(
-                tenant == null ? (int?)null : (int?)tenant.Id,
+                tenant == null ? (long?)null : (long?)tenant.Id,
                 tenant == null ? MultiTenancySides.Host : MultiTenancySides.Tenant
                 );
