+++ src/Abp.ZeroCore/Authorization/Users/AbpUserManager.cs
@@ -428,22 +428,22 @@ namespace Abp.Authorization.Users
             return AbpUserStore.FindAll(login);
         }
 
-        public virtual Task<TUser> FindAsync(int? tenantId, UserLoginInfo login)
+        public virtual Task<TUser> FindAsync(long? tenantId, UserLoginInfo login)
         {
             return AbpUserStore.FindAsync(tenantId, login);
         }
 
-        public virtual TUser Find(int? tenantId, UserLoginInfo login)
+        public virtual TUser Find(long? tenantId, UserLoginInfo login)
         {
             return AbpUserStore.Find(tenantId, login);
         }
 
-        public virtual Task<TUser> FindByNameOrEmailAsync(int? tenantId, string userNameOrEmailAddress)
+        public virtual Task<TUser> FindByNameOrEmailAsync(long? tenantId, string userNameOrEmailAddress)
         {
             return AbpUserStore.FindByNameOrEmailAsync(tenantId, userNameOrEmailAddress);
         }
 
-        public virtual TUser FindByNameOrEmail(int? tenantId, string userNameOrEmailAddress)
+        public virtual TUser FindByNameOrEmail(long? tenantId, string userNameOrEmailAddress)
         {
             return AbpUserStore.FindByNameOrEmail(tenantId, userNameOrEmailAddress);
         }
@@ -796,7 +796,7 @@ namespace Abp.Authorization.Users
                 );
         }
 
-        private async Task CheckMaxUserOrganizationUnitMembershipCountAsync(int? tenantId, int requestedCount)
+        private async Task CheckMaxUserOrganizationUnitMembershipCountAsync(long? tenantId, int requestedCount)
         {
             var maxCount = await _organizationUnitSettings.GetMaxUserMembershipCountAsync(tenantId);
             if (requestedCount > maxCount)
@@ -805,7 +805,7 @@ namespace Abp.Authorization.Users
             }
         }
 
-        private void CheckMaxUserOrganizationUnitMembershipCount(int? tenantId, int requestedCount)
+        private void CheckMaxUserOrganizationUnitMembershipCount(long? tenantId, int requestedCount)
         {
             var maxCount = _organizationUnitSettings.GetMaxUserMembershipCount(tenantId);
             if (requestedCount > maxCount)
@@ -953,7 +953,7 @@ namespace Abp.Authorization.Users
             }
         }
 
-        public virtual async Task InitializeOptionsAsync(int? tenantId)
+        public virtual async Task InitializeOptionsAsync(long? tenantId)
         {
             Options = JsonConvert.DeserializeObject<IdentityOptions>(_optionsAccessor.Value.ToJsonString());
 
@@ -970,7 +970,7 @@ namespace Abp.Authorization.Users
             Options.Password.RequiredLength = await GetSettingValueAsync<int>(AbpZeroSettingNames.UserManagement.PasswordComplexity.RequiredLength, tenantId);
         }
 
-        public virtual void InitializeOptions(int? tenantId)
+        public virtual void InitializeOptions(long? tenantId)
         {
             Options = JsonConvert.DeserializeObject<IdentityOptions>(_optionsAccessor.Value.ToJsonString());
 
@@ -1114,24 +1114,24 @@ namespace Abp.Authorization.Users
         //    return providers;
         //}
 
-        private bool IsTrue(string settingName, int? tenantId)
+        private bool IsTrue(string settingName, long? tenantId)
         {
             return GetSettingValue<bool>(settingName, tenantId);
         }
 
-        private Task<bool> IsTrueAsync(string settingName, int? tenantId)
+        private Task<bool> IsTrueAsync(string settingName, long? tenantId)
         {
             return GetSettingValueAsync<bool>(settingName, tenantId);
         }
 
-        private T GetSettingValue<T>(string settingName, int? tenantId) where T : struct
+        private T GetSettingValue<T>(string settingName, long? tenantId) where T : struct
         {
             return tenantId == null
                 ? _settingManager.GetSettingValueForApplication<T>(settingName)
                 : _settingManager.GetSettingValueForTenant<T>(settingName, tenantId.Value);
         }
 
-        private Task<T> GetSettingValueAsync<T>(string settingName, int? tenantId) where T : struct
+        private Task<T> GetSettingValueAsync<T>(string settingName, long? tenantId) where T : struct
         {
             return tenantId == null
                 ? _settingManager.GetSettingValueForApplicationAsync<T>(settingName)
@@ -1148,7 +1148,7 @@ namespace Abp.Authorization.Users
             return LocalizationManager.GetString(LocalizationSourceName, name, cultureInfo);
         }
 
-        private int? GetCurrentTenantId()
+        private long? GetCurrentTenantId()
         {
             if (_unitOfWorkManager.Current != null)
             {
